+++ tests/test_printprogress.py
@@ -3,7 +3,7 @@
 """Tests for `lognflow` package."""
 
 import pytest
-
+import inspect
 from lognflow import lognflow, select_directory, logviewer, printprogress
 
 import numpy as np
@@ -13,6 +13,7 @@ import tempfile
 temp_dir = tempfile.gettempdir()
 
 def test_printprogress():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     for N in list([100, 200, 400, 1000]):
         pprog = printprogress(N)
         for _ in range(N):
@@ -20,6 +21,7 @@ def test_printprogress():
             pprog()
 
 def test_printprogress_with_logger():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     logger = lognflow(temp_dir)
     N = 1500000
     pprog = printprogress(N, print_function = logger, log_time_stamp = False)
@@ -27,6 +29,7 @@ def test_printprogress_with_logger():
         pprog()
         
 def test_printprogress_ETA():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     logger = lognflow(temp_dir)
     N = 500000
     pprog = printprogress(N, print_function = None)
@@ -35,6 +38,7 @@ def test_printprogress_ETA():
         print(f'ETA: {ETA:.2f}')
     
 def test_specific_timing():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     logger = lognflow(temp_dir)
     N = 7812
     pprog = printprogress(N, title='Inference of 7812 points. ')
@@ -45,6 +49,7 @@ def test_specific_timing():
         pprog()
 
 def test_generator_type():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     vec = np.arange(12)
     sum = 0
     for _ in printprogress(vec):
@@ -53,6 +58,7 @@ def test_generator_type():
     print(f'sum: {sum}')
 
 def test_varying_periods():
+    print('Testing function', inspect.currentframe().f_code.co_name)
     vec = np.arange(30)
     sum = 0
     for _ in printprogress(vec):